Update artifact workflows, added macOS artifact creation
Updated workflows to better handle artifact uploads. This includes better cleanup of no longer dumping all the possible platforms into the main directory of the CI job, as well as for each platform type, specific artifacts for each platform is made to lower artifact sizes and to be specific to the correct platform.

Added other adjustments for platform types across the workflow for future ARM-specific builds

Added macOS artifact creation.