Fix compiler issues and broken functions
Adding more compiler args to prevent other platforms that are not able to compile some platforms to compile files, as it would just error and fail the build.

Fixing broken functions such as `SocketCANDevice::GetID()` as it was not returning a string. This is now not implemented but it returns an `int` which is correct for the function.